FIX: View setup error in custom configuration Error was apparent in cave-emu configuration: views were missing output. Error was caused by tiles being disabled due wrong view / tile viewport intersection, caused by the fact that custom configurations did not have a computed global canvas pixel size.
Now the display configuration always computes a global canvas pixel size
based on the maximum pixel viewport extent of all the tiles present in
the configuration.

A somewhat related fix in the Camera class, use the tile pixel size as
the canvas size to check if drawing is enabled in a custom camera tile.